fix: pr notes + minor refactoring
