+++ tests/mockData/testStaticData.ts
@@ -2,10 +2,10 @@
 import { IJobResponse, ITaskResponse, OperationStatus } from '@map-colonies/mc-priority-queue';
 import { ITaskParams } from '../../src/common/interfaces';
 import { CacheType, SeedMode } from '../../src/common/enums';
-// import * as zx from 'zx';
-// import { ProcessOutput, ProcessPromise, $ } from 'zx';
-// import { Readable, Writable } from 'stream';
-// import { ChildProcess } from 'child_process';
+import { Readable, Writable } from 'stream';
+import { ProcessOutput, ProcessPromise } from 'zx';
+import { ChildProcess } from 'child_process';
+import * as zx from 'zx';
 
 const task = {
   id: '9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d',
@@ -103,39 +103,39 @@ const job: IJobResponse<unknown, unknown> = {
   created: '2024-01-16T16:21:53.020Z',
   updated: '2024-01-16T16:21:53.020Z',
 };
-//   updated: '2024-01-16T16:21:53.020Z',
+
 function getJob(): IJobResponse<unknown, unknown> {
   return job;
 }
 
-// const readable = Readable.from(['some test data', 'some test data2']);
-// const cmdProcessPromise: ProcessPromise<ProcessOutput> = {
-//   stdout: readable,
-//   child: new ChildProcess(),
-//   stdin: new Writable(),
-//   stderr: new Readable(),
-//   exitCode: Promise.resolve(0),
-//   pipe: function (dest: zx.ProcessPromise<zx.ProcessOutput> | Writable): zx.ProcessPromise<zx.ProcessOutput> {
-//     throw new Error('Function not implemented.');
-//   },
-//   kill: function (signal?: string | number | undefined): Promise<void> {
-//     throw new Error('Function not implemented.');
-//   },
-//   then: function <TResult1 = zx.ProcessOutput, TResult2 = never>(
-//     onfulfilled?: ((value: zx.ProcessOutput) => TResult1 | PromiseLike<TResult1>) | null | undefined,
-//     onrejected?: ((reason: any) => TResult2 | PromiseLike<TResult2>) | null | undefined
-//   ): Promise<TResult1 | TResult2> {
-//     throw new Error('Function not implemented.');
-//   },
-//   catch: function <TResult = never>(
-//     onrejected?: ((reason: any) => TResult | PromiseLike<TResult>) | null | undefined
-//   ): Promise<zx.ProcessOutput | TResult> {
-//     throw new Error('Function not implemented.');
-//   },
-//   finally: function (onfinally?: (() => void) | null | undefined): Promise<zx.ProcessOutput> {
-//     throw new Error('Function not implemented.');
-//   },
-//   [Symbol.toStringTag]: '',
-// };
-// export { getTask, getJob, cmdProcessPromise };
-export { getTask, getJob };
+const readable = Readable.from(['some test data', 'some test data2']);
+const cmdProcessPromise: ProcessPromise<ProcessOutput> = {
+  stdout: readable,
+  child: new ChildProcess(),
+  stdin: new Writable(),
+  stderr: new Readable(),
+  exitCode: Promise.resolve(0),
+  pipe: function (dest: zx.ProcessPromise<zx.ProcessOutput> | Writable): zx.ProcessPromise<zx.ProcessOutput> {
+    throw new Error('Function not implemented.');
+  },
+  kill: function (signal?: string | number | undefined): Promise<void> {
+    throw new Error('Function not implemented.');
+  },
+  then: function <TResult1 = zx.ProcessOutput, TResult2 = never>(
+    onfulfilled?: ((value: zx.ProcessOutput) => TResult1 | PromiseLike<TResult1>) | null | undefined,
+    onrejected?: ((reason: any) => TResult2 | PromiseLike<TResult2>) | null | undefined
+  ): Promise<TResult1 | TResult2> {
+    throw new Error('Function not implemented.');
+  },
+  catch: function <TResult = never>(
+    onrejected?: ((reason: any) => TResult | PromiseLike<TResult>) | null | undefined
+  ): Promise<zx.ProcessOutput | TResult> {
+    throw new Error('Function not implemented.');
+  },
+  finally: function (onfinally?: (() => void) | null | undefined): Promise<zx.ProcessOutput> {
+    throw new Error('Function not implemented.');
+  },
+  [Symbol.toStringTag]: '',
+};
+
+export { getTask, getJob, cmdProcessPromise };
